(refactor): made BooksList as another component
