+++ react/finger-exercises/redux/02 - react-redux/src/app/screens/App/components/BooksShop/index.js
@@ -3,63 +3,23 @@ import { connect } from 'react-redux';
 import Navbar from '@components/Navbar';
 import Footer from '@components/Footer';
 
-import { DATA as fakeBooks } from '../../../../../constants/data';
-
-import Book from './../Book';
+import BooksList from './../BooksList';
 import Search from './../Search';
-import ShoppingCart from './../ShoppingCart';
 import styles from './styles.scss';
 
 class BooksShop extends Component {
-  state = {
-    books: fakeBooks,
-    bookSelected: []
-  };
-
-  CONFIGURATION_BUTTON = {
-    add: {
-      text: 'Add to cart',
-      function: this.addToCart
-    },
-    remove: {
-      text: 'Remove',
-      function: this.removeItem,
-      isDanger: true
-    }
-  };
-
-  renderBooks = item => {
-    const showButton = !this.state.bookSelected.some(el => el.id === item.id);
-    const configButton = showButton ? this.CONFIGURATION_BUTTON.add : this.CONFIGURATION_BUTTON.remove;
-    return <Book key={item.id} data={item} configButton={configButton} />;
-  };
-
   render() {
     return (
       <Fragment>
         <Navbar />
         <div className={styles.container}>
           <Search onSearch={this.onSearch} />
-          {this.state.books.length ? (
-            this.state.books.map(this.renderBooks)
-          ) : (
-            <div className={styles.noData}>
-              <h2 className={styles.title}>No Data</h2>
-            </div>
-          )}
+          <BooksList />
         </div>
-        {this.state.bookSelected.length ? (
-          <ShoppingCart data={this.state.bookSelected} addItem={this.addItem} removeItem={this.removeItem} />
-        ) : null}
         <Footer />
       </Fragment>
     );
   }
 }
 
-const mapStateToProps = state => ({
-  books: state.books,
-  originalBooks: state.originalBooks
-});
-
-export default connect(mapStateToProps)(BooksShop);
+export default connect()(BooksShop);
